+++ tests/scheduler/backends/test_zhinst_backend.py
@@ -911,7 +911,7 @@ def test_uhfqa_sequence3(
         ro_pulse_frequency=7.04e9,
         ro_acquisition_delay=ro_acquisition_delay,
         ro_integration_time=500e-9,
-        buffer_time=1e-5,
+        init_duration=1e-5,
     )
     schedule = create_schedule_with_pulse_info(schedule)